Relax some rules
The "no-null" is too big of a change for our code base. Causes lots and lots of errors all over the place + changing from null to undefined can be breaking as "null !== undefined", so any current type safe check against null or undefined could end up becoming wrong.

The other two relaxations is to make it easier to upgrade.